+++ example/pages/dashboard/config.js
@@ -1,0 +1,86 @@
+export default {
+  base: {
+    title: '基础组件',
+    content: [
+      {
+        name: 'Badge 徽章',
+        path: '/pages/badge/index'
+      }, {
+        name: 'Button 按钮',
+        path: '/pages/btn/index'
+      }, {
+        name: 'Capsule 胶囊',
+        path: '/pages/capsule/index'
+      }, {
+        name: 'Card 卡片',
+        path: '/pages/card/index'
+      }, {
+        name: 'Cell 单元格',
+        path: '/pages/cell/index'
+      }, {
+        name: 'Helper 基础样式',
+        path: '/pages/helper/index'
+      }, {
+        name: 'Icon 图标',
+        path: '/pages/icon/index'
+      }, {
+        name: 'Label 标签',
+        path: '/pages/label/index'
+      }, {
+        name: 'Layout 标签',
+        path: '/pages/layout/index'
+      }, {
+        name: 'Loadmore 加载',
+        path: '/pages/loadmore/index'
+      }, {
+        name: 'Noticebar 通告栏',
+        path: '/pages/noticebar/index'
+      }, {
+        name: 'Panel 面板',
+        path: '/pages/panel/index'
+      }, {
+        name: 'Popup 弹出层',
+        path: '/pages/popup/index'
+      }, {
+        name: 'Quantity 计数器',
+        path: '/pages/quantity/index'
+      }, {
+        name: 'Select 选择',
+        path: '/pages/select/index'
+      }, {
+        name: 'Steps 步骤条',
+        path: '/pages/steps/index'
+      }, {
+        name: 'Switch 开关',
+        path: '/pages/switch/index'
+      }, {
+        name: 'Tab 标签',
+        path: '/pages/tab/index'
+      }
+    ]
+  },
+  form: {
+    title: '表单',
+    content: [
+      {
+        name: 'Field 输入框',
+        path: '/pages/field/index'
+      }
+    ]
+  },
+  action: {
+    title: '操作反馈',
+    content: [
+      {
+        name: 'Dialog 弹出框',
+        path: '/pages/dialog/index'
+      }, {
+        name: 'Toast 轻提示',
+        path: '/pages/toast/index'
+      }, {
+        name: 'Toptips 顶部提示',
+        path: '/pages/toptips/index'
+      }
+    ]
+  }
+};