+++ example/pages/dashboard/index.js
@@ -1,0 +1,14 @@
+import componentsConfig from './config';
+
+Page({
+  data: {
+    list: componentsConfig
+  },
+
+  onLoad: function () {
+
+  },
+
+  onShow: function() {
+  },
+})