+++ example/pages/panel/index.js
@@ -1,0 +1,11 @@
+Page({
+  data: {
+  },
+
+  onLoad: function () {
+
+  },
+
+  onShow: function() {
+  },
+})