+++ changelog.md
@@ -1,4 +1,47 @@
 `ZanUI-WeApp` 严格遵循 [Semantic Versioning 2.0.0](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
+## [v2.3.2](https://github.com/youzan/zanui-weapp/tree/v2.3.2) (2017-11-02)
+[Full Changelog](https://github.com/youzan/zanui-weapp/compare/v2.3.1...v2.3.2)
+
+**Improvements**
+
+- 更新 Icon 中 shop 图标 [\#45](https://github.com/youzan/zanui-weapp/pull/45) ([pangxie1991](https://github.com/pangxie1991))
+- 新增 layout 组件 [\#43](https://github.com/youzan/zanui-weapp/pull/43) ([pangxie1991](https://github.com/pangxie1991))
+- Field 组件中 Input 事件支持升级 [\#42](https://github.com/youzan/zanui-weapp/pull/42) ([pangxie1991](https://github.com/pangxie1991))
+- 增加 css 编译，方便组件书写 [\#40](https://github.com/youzan/zanui-weapp/pull/40) ([pangxie1991](https://github.com/pangxie1991))
+
+## [v2.3.1](https://github.com/youzan/zanui-weapp/tree/v2.3.1) (2017-10-27)
+[Full Changelog](https://github.com/youzan/zanui-weapp/compare/v2.3.0...v2.3.1)
+
+**Bug Fixes**
+
+- 修复 Field 组件多行展示样式问题 [\#39](https://github.com/youzan/zanui-weapp/pull/39) ([pangxie1991](https://github.com/pangxie1991))
+
+## [v2.3.0](https://github.com/youzan/zanui-weapp/tree/v2.3.0) (2017-10-26)
+[Full Changelog](https://github.com/youzan/zanui-weapp/compare/v2.2.0...v2.3.0)
+
+**Breaking Change**
+
+- icon 整理 & 新增 [\#38](https://github.com/youzan/zanui-weapp/pull/38) ([pangxie1991](https://github.com/pangxie1991))
+- 新增 Field 组件, 代替原有 Form 组件 [\#37](https://github.com/youzan/zanui-weapp/pull/37) ([pangxie1991](https://github.com/pangxie1991))
+- 新增 Select 组件，代替原有 Check_label 组件 [\#36](https://github.com/youzan/zanui-weapp/pull/36) ([pangxie1991](https://github.com/pangxie1991))
+
+## [v2.2.0](https://github.com/youzan/zanui-weapp/tree/v2.2.0) (2017-10-15)
+[Full Changelog](https://github.com/youzan/zanui-weapp/compare/v2.1.1...v2.2.0)
+
+**Bug Fixes**
+
+- 多页面情况下，TopTips 无法自动清除 [\#20](https://github.com/youzan/zanui-weapp/issues/20)
+
+**Improvements**
+
+- \[new feature\] 增加发布脚本 [\#35](https://github.com/youzan/zanui-weapp/pull/35) ([pangxie1991](https://github.com/pangxie1991))
+- \[new feature\] 增加Capsule, Noticebar, Popup；修正Dialog在非最外层情况下引入的样式问题；更新readme。 [\#32](https://github.com/youzan/zanui-weapp/pull/32) ([everywill](https://github.com/everywill))
+- \[new feature\] 增加单选 和一个验证码的样式 [\#30](https://github.com/youzan/zanui-weapp/pull/30) ([muzea](https://github.com/muzea))
+- \[new feature\] Docs: 文档更新 [\#29](https://github.com/youzan/zanui-weapp/pull/29) ([pangxie1991](https://github.com/pangxie1991))
+
 ## 2.1.0
+
+**Improvements**
+
 - 新增`switch`组件